+++ tests/testthat/test-covlmc_draw.R
@@ -5,7 +5,7 @@ test_that("draw obeys its contract (with vgam)", {
   y <- as.factor(ifelse(runif(length(dts)) > 0.2, y, sample(c("A", "B", "C"), length(dts), replace = TRUE)))
   df_y <- data.frame(y = y)
   model <- covlmc(dts, df_y, alpha = 0.01, min_size = 1.5)
-  expect_snapshot_output(draw(model, model = NULL, p_value = FALSE))
+  expect_snapshot_output(draw(model, model = NULL, p_value = TRUE))
   expect_snapshot_output(draw(prune(model, 0.0001)))
   expect_snapshot_output(draw(prune(model, 0.0001), time_sep = " | "))
   expect_snapshot_output(draw(prune(model, 0.0001), model = "full", time_sep = " | "))